refactor: various generator code refactorings and cleanups
