+++ packages/client-generator-ts/src/TSClient/Datasources.ts
@@ -1,14 +1,17 @@
 import { DataSource } from '@prisma/generator'
 import indent from 'indent-string'
 
-import type { Generable } from './Generable'
-
-export class Datasources implements Generable {
+export class Datasources {
   constructor(protected readonly internalDatasources: DataSource[]) {}
   public toTS(): string {
     const sources = this.internalDatasources
-    return `export type Datasources = {
+    return `
+export type Datasource = {
+  url?: string
+}
+export type Datasources = {
 ${indent(sources.map((s) => `${s.name}?: Datasource`).join('\n'), 2)}
-}`
+}
+`
   }
 }
