+++ packages/client-generator-ts/src/TSClient/FieldRefInput.ts
@@ -1,9 +1,8 @@
 import type * as DMMF from '@prisma/dmmf'
 
 import { getRefAllowedTypeName } from '../utils'
-import type { Generable } from './Generable'
 
-export class FieldRefInput implements Generable {
+export class FieldRefInput {
   constructor(private type: DMMF.FieldRefType) {}
 
   toTS() {
