--- packages/client-generator-ts/src/TSClient/Generable.ts
@@ -1,3 +1,0 @@
-export interface Generable {
-  toTS(): string
-}
