+++ packages/client-generator-ts/src/TSClient/GenerateContext.ts
@@ -1,22 +1,39 @@
 import { GeneratorConfig } from '@prisma/generator'
 
 import { DMMFHelper } from '../dmmf'
+import { FileNameMapper } from '../file-extensions'
 import { GenericArgsInfo } from '../GenericsArgsInfo'
 
 export interface GenerateContextOptions {
   dmmf: DMMFHelper
   genericArgsInfo: GenericArgsInfo
+  runtimeImport: string
+  outputFileName: FileNameMapper
+  importFileName: FileNameMapper
   generator?: GeneratorConfig
 }
 
 export class GenerateContext implements GenerateContextOptions {
   dmmf: DMMFHelper
   genericArgsInfo: GenericArgsInfo
+  runtimeImport: string
+  outputFileName: FileNameMapper
+  importFileName: FileNameMapper
   generator?: GeneratorConfig
 
-  constructor({ dmmf, genericArgsInfo, generator }: GenerateContextOptions) {
+  constructor({
+    dmmf,
+    genericArgsInfo,
+    runtimeImport,
+    outputFileName,
+    importFileName,
+    generator,
+  }: GenerateContextOptions) {
     this.dmmf = dmmf
     this.genericArgsInfo = genericArgsInfo
+    this.runtimeImport = runtimeImport
+    this.outputFileName = outputFileName
+    this.importFileName = importFileName
     this.generator = generator
   }
