+++ packages/client-generator-ts/src/TSClient/Input.ts
@@ -7,10 +7,9 @@ import { GenericArgsInfo } from '../GenericsArgsInfo'
 import { appendSkipType } from '../utils'
 import { GraphQLScalarToJSTypeTable, JSOutputTypeToInputType } from '../utils/common'
 import { TAB_SIZE } from './constants'
-import type { Generable } from './Generable'
 import { GenerateContext } from './GenerateContext'
 
-export class InputField implements Generable {
+export class InputField {
   constructor(
     protected readonly field: DMMF.SchemaArg,
     protected readonly context: GenerateContext,
@@ -123,7 +122,7 @@ function xorTypes(types: ts.TypeBuilder[]) {
   return types.reduce((prev, curr) => ts.namedType('Prisma.XOR').addGenericArgument(prev).addGenericArgument(curr))
 }
 
-export class InputType implements Generable {
+export class InputType {
   private generatedName: string
   constructor(protected readonly type: DMMF.InputType, protected readonly context: GenerateContext) {
     this.generatedName = type.name
