+++ packages/client-generator-ts/src/TSClient/PrismaClient.ts
@@ -1,179 +1,14 @@
-import { lowerCase, NonModelOperation, Operation } from '@prisma/client-common'
+import { lowerCase } from '@prisma/client-common'
 import type * as DMMF from '@prisma/dmmf'
-import type { DataSource } from '@prisma/generator'
-import { assertNever } from '@prisma/internals'
 import * as ts from '@prisma/ts-builders'
 import indent from 'indent-string'
 
-import {
-  capitalize,
-  extArgsParam,
-  getAggregateName,
-  getCountAggregateOutputName,
-  getFieldRefsTypeName,
-  getGroupByName,
-  getModelArgName,
-  getPayloadName,
-} from '../utils'
+import { capitalize } from '../utils'
 import { runtimeImport, runtimeImportedType } from '../utils/runtimeImport'
 import { TAB_SIZE } from './constants'
-import { Datasources } from './Datasources'
-import type { Generable } from './Generable'
 import { GenerateContext } from './GenerateContext'
-import { globalOmitConfig } from './globalOmit'
 import { TSClientOptions } from './TSClient'
-import { getModelActions } from './utils/getModelActions'
 import * as tsx from './utils/type-builders'
-
-function clientTypeMapModelsDefinition(context: GenerateContext) {
-  const meta = ts.objectType()
-
-  const modelNames = context.dmmf.datamodel.models.map((m) => m.name)
-
-  // `modelNames` can be empty if `generate --allow-no-models` is used.
-  if (modelNames.length === 0) {
-    meta.add(ts.property('modelProps', ts.neverType))
-  } else {
-    meta.add(ts.property('modelProps', ts.unionType(modelNames.map((name) => ts.stringLiteral(lowerCase(name))))))
-  }
-
-  const isolationLevel = context.dmmf.hasEnumInNamespace('TransactionIsolationLevel', 'prisma')
-    ? ts.namedType('TransactionIsolationLevel')
-    : ts.neverType
-  meta.add(ts.property('txIsolationLevel', isolationLevel))
-
-  const model = ts.objectType()
-
-  model.addMultiple(
-    modelNames.map((modelName) => {
-      const entry = ts.objectType()
-      entry.add(
-        ts.property('payload', ts.namedType(getPayloadName(modelName)).addGenericArgument(extArgsParam.toArgument())),
-      )
-      entry.add(ts.property('fields', ts.namedType(`Prisma.${getFieldRefsTypeName(modelName)}`)))
-      const actions = getModelActions(context.dmmf, modelName)
-      const operations = ts.objectType()
-      operations.addMultiple(
-        actions.map((action) => {
-          const operationType = ts.objectType()
-          const argsType = `Prisma.${getModelArgName(modelName, action)}`
-          operationType.add(ts.property('args', ts.namedType(argsType).addGenericArgument(extArgsParam.toArgument())))
-          operationType.add(ts.property('result', clientTypeMapModelsResultDefinition(modelName, action)))
-          return ts.property(action, operationType)
-        }),
-      )
-      entry.add(ts.property('operations', operations))
-      return ts.property(modelName, entry)
-    }),
-  )
-
-  return ts
-    .objectType()
-    .add(ts.property('globalOmitOptions', ts.objectType().add(ts.property('omit', ts.namedType('GlobalOmitOptions')))))
-    .add(ts.property('meta', meta))
-    .add(ts.property('model', model))
-}
-
-function clientTypeMapModelsResultDefinition(
-  modelName: string,
-  action: Exclude<Operation, `$${string}`>,
-): ts.TypeBuilder {
-  if (action === 'count')
-    return ts.unionType([tsx.optional(ts.namedType(`Prisma.${getCountAggregateOutputName(modelName)}`)), ts.numberType])
-  if (action === 'groupBy') return ts.array(tsx.optional(ts.namedType(`Prisma.${getGroupByName(modelName)}`)))
-  if (action === 'aggregate') return tsx.optional(ts.namedType(`Prisma.${getAggregateName(modelName)}`))
-  if (action === 'findRaw') return ts.namedType('Prisma.JsonObject')
-  if (action === 'aggregateRaw') return ts.namedType('Prisma.JsonObject')
-  if (action === 'deleteMany') return ts.namedType('BatchPayload')
-  if (action === 'createMany') return ts.namedType('BatchPayload')
-  if (action === 'createManyAndReturn') return ts.array(payloadToResult(modelName))
-  if (action === 'updateMany') return ts.namedType('BatchPayload')
-  if (action === 'updateManyAndReturn') return ts.array(payloadToResult(modelName))
-  if (action === 'findMany') return ts.array(payloadToResult(modelName))
-  if (action === 'findFirst') return ts.unionType([payloadToResult(modelName), ts.nullType])
-  if (action === 'findUnique') return ts.unionType([payloadToResult(modelName), ts.nullType])
-  if (action === 'findFirstOrThrow') return payloadToResult(modelName)
-  if (action === 'findUniqueOrThrow') return payloadToResult(modelName)
-  if (action === 'create') return payloadToResult(modelName)
-  if (action === 'update') return payloadToResult(modelName)
-  if (action === 'upsert') return payloadToResult(modelName)
-  if (action === 'delete') return payloadToResult(modelName)
-
-  assertNever(action, `Unknown action: ${action}`)
-}
-
-function payloadToResult(modelName: string) {
-  return ts.namedType('runtime.Types.Utils.PayloadToResult').addGenericArgument(ts.namedType(getPayloadName(modelName)))
-}
-
-function clientTypeMapOthersDefinition(context: GenerateContext) {
-  const otherOperationsNames = context.dmmf.getOtherOperationNames().flatMap((name) => {
-    const results = [`$${name}`]
-    if (name === 'executeRaw' || name === 'queryRaw') {
-      results.push(`$${name}Unsafe`)
-    }
-
-    if (name === 'queryRaw' && context.isPreviewFeatureOn('typedSql')) {
-      results.push(`$queryRawTyped`)
-    }
-
-    return results
-  })
-
-  const argsResultMap = {
-    $executeRaw: { args: '[query: TemplateStringsArray | Sql, ...values: any[]]', result: 'any' },
-    $queryRaw: { args: '[query: TemplateStringsArray | Sql, ...values: any[]]', result: 'any' },
-    $executeRawUnsafe: { args: '[query: string, ...values: any[]]', result: 'any' },
-    $queryRawUnsafe: { args: '[query: string, ...values: any[]]', result: 'any' },
-    $runCommandRaw: { args: 'Prisma.InputJsonObject', result: 'JsonObject' },
-    $queryRawTyped: { args: 'runtime.UnknownTypedSql', result: 'JsonObject' },
-  } satisfies Record<NonModelOperation, { args: string; result: string }>
-
-  return `{
-  other: {
-    payload: any
-    operations: {${otherOperationsNames.reduce((acc, action) => {
-      return `${acc}
-      ${action}: {
-        args: ${argsResultMap[action].args},
-        result: ${argsResultMap[action].result}
-      }`
-    }, '')}
-    }
-  }
-}`
-}
-
-function clientTypeMapDefinition(context: GenerateContext) {
-  const typeMap = `${ts.stringify(clientTypeMapModelsDefinition(context))} & ${clientTypeMapOthersDefinition(context)}`
-
-  return `
-export interface TypeMapCb<ClientOptions = {}> extends runtime.Types.Utils.Fn<{extArgs: runtime.Types.Extensions.InternalArgs }, runtime.Types.Utils.Record<string, any>> {
-  returns: TypeMap<this['params']['extArgs'], ClientOptions extends { omit: infer OmitOptions } ? OmitOptions : {}>
-}
-
-export type TypeMap<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs, GlobalOmitOptions = {}> = ${typeMap}`
-}
-
-function clientExtensionsDefinitions(context: GenerateContext) {
-  const typeMap = clientTypeMapDefinition(context)
-  const define = ts.moduleExport(
-    ts.constDeclaration('defineExtension').setValue(
-      ts
-        .namedValue('runtime.Extensions.defineExtension')
-        .as(ts.namedType('unknown'))
-        .as(
-          ts
-            .namedType('runtime.Types.Extensions.ExtendsHook')
-            .addGenericArgument(ts.stringLiteral('define'))
-            .addGenericArgument(ts.namedType('TypeMapCb'))
-            .addGenericArgument(ts.namedType('runtime.Types.Extensions.DefaultArgs')),
-        ),
-    ),
-  )
-
-  return [typeMap, ts.stringify(define)].join('\n')
-}
 
 function extendsPropertyDefinition() {
   const extendsDefinition = ts
@@ -402,21 +237,6 @@ function runCommandRawDefinition(context: GenerateContext) {
   return ts.stringify(method, { indentLevel: 1, newLine: 'leading' })
 }
 
-function applyPendingMigrationsDefinition(this: PrismaClientClass) {
-  if (this.runtimeName !== 'react-native') {
-    return null
-  }
-
-  const method = ts
-    .method('$applyPendingMigrations')
-    .setReturnType(tsx.promise(ts.voidType))
-    .setDocComment(
-      ts.docComment`Tries to apply pending migrations one by one. If a migration fails to apply, the function will stop and throw an error. You are responsible for informing the user and possibly blocking the app as we cannot guarantee the state of the database.`,
-    )
-
-  return ts.stringify(method, { indentLevel: 1, newLine: 'leading' })
-}
-
 function eventRegistrationMethodDeclaration(runtimeName: TSClientOptions['runtimeName']) {
   if (runtimeName === 'binary') {
     return `$on<V extends (U | 'beforeExit')>(eventType: V, callback: (event: V extends 'query' ? Prisma.QueryEvent : V extends 'beforeExit' ? () => runtime.Types.Utils.JsPromise<void> : Prisma.LogEvent) => void): PrismaClient;`
@@ -425,13 +245,12 @@ function eventRegistrationMethodDeclaration(runtimeName: TSClientOptions['runtim
   }
 }
 
-export class PrismaClientClass implements Generable {
+export class PrismaClientClass {
   constructor(
-    protected readonly context: GenerateContext,
-    protected readonly internalDatasources: DataSource[],
-    protected readonly outputDir: string,
-    protected readonly runtimeName: TSClientOptions['runtimeName'],
+    private readonly context: GenerateContext,
+    private readonly runtimeName: TSClientOptions['runtimeName'],
   ) {}
+
   private get jsDoc(): string {
     const { dmmf } = this.context
 
@@ -461,7 +280,8 @@ export class PrismaClientClass implements Generable {
  * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client).
  */`
   }
-  public toTSWithoutNamespace(): string {
+
+  public toTS(): string {
     const { dmmf } = this.context
 
     return `\
@@ -509,7 +329,7 @@ ${[
   interactiveTransactionDefinition(this.context),
   runCommandRawDefinition(this.context),
   metricDefinition(this.context),
-  applyPendingMigrationsDefinition.bind(this)(),
+  this.applyPendingMigrationsDefinition(),
   extendsPropertyDefinition(),
 ]
   .filter((d) => d !== null)
@@ -541,181 +361,22 @@ get ${methodName}(): Prisma.${m.model}Delegate<${generics.join(', ')}>;`
     )}
 }
 
-export const PrismaClient = runtime.getPrismaClient(config) as unknown as PrismaClientConstructor`
-  }
-  public toTS(): string {
-    const clientOptions = this.buildClientOptions()
-
-    return `${new Datasources(this.internalDatasources).toTS()}
-${clientExtensionsDefinitions(this.context)}
-export type DefaultPrismaClient = PrismaClient
-export type ErrorFormat = 'pretty' | 'colorless' | 'minimal'
-${ts.stringify(ts.moduleExport(clientOptions))}
-${ts.stringify(globalOmitConfig(this.context.dmmf))}
-
-/* Types for Logging */
-export type LogLevel = 'info' | 'query' | 'warn' | 'error'
-export type LogDefinition = {
-  level: LogLevel
-  emit: 'stdout' | 'event'
-}
-
-export type GetLogType<T extends LogLevel | LogDefinition> = T extends LogDefinition ? T['emit'] extends 'event' ? T['level'] : never : never
-export type GetEvents<T extends any> = T extends Array<LogLevel | LogDefinition> ?
-  GetLogType<T[0]> | GetLogType<T[1]> | GetLogType<T[2]> | GetLogType<T[3]>
-  : never
-
-export type QueryEvent = {
-  timestamp: Date
-  query: string
-  params: string
-  duration: number
-  target: string
-}
-
-export type LogEvent = {
-  timestamp: Date
-  message: string
-  target: string
-}
-/* End Types for Logging */
-
-
-export type PrismaAction =
-  | 'findUnique'
-  | 'findUniqueOrThrow'
-  | 'findMany'
-  | 'findFirst'
-  | 'findFirstOrThrow'
-  | 'create'
-  | 'createMany'
-  | 'createManyAndReturn'
-  | 'update'
-  | 'updateMany'
-  | 'updateManyAndReturn'
-  | 'upsert'
-  | 'delete'
-  | 'deleteMany'
-  | 'executeRaw'
-  | 'queryRaw'
-  | 'aggregate'
-  | 'count'
-  | 'runCommandRaw'
-  | 'findRaw' 
-  | 'groupBy'
-
-/**
- * These options are being passed into the middleware as "params"
- */
-export type MiddlewareParams = {
-  model?: ModelName
-  action: PrismaAction
-  args: any
-  dataPath: string[]
-  runInTransaction: boolean
-}
-
-/**
- * The \`T\` type makes sure, that the \`return proceed\` is not forgotten in the middleware implementation
- */
-export type Middleware<T = any> = (
-  params: MiddlewareParams,
-  next: (params: MiddlewareParams) => runtime.Types.Utils.JsPromise<T>,
-) => runtime.Types.Utils.JsPromise<T>
-
-/**
- * \`PrismaClient\` proxy available in interactive transactions.
- */
-export type TransactionClient = Omit<DefaultPrismaClient, runtime.ITXClientDenyList>
+export const PrismaClient = runtime.getPrismaClient(config) as unknown as PrismaClientConstructor
 `
   }
 
-  private buildClientOptions() {
-    const clientOptions = ts
-      .interfaceDeclaration('PrismaClientOptions')
-      .add(
-        ts
-          .property('datasources', ts.namedType('Datasources'))
-          .optional()
-          .setDocComment(ts.docComment('Overwrites the datasource url from your schema.prisma file')),
-      )
-      .add(
-        ts
-          .property('datasourceUrl', ts.stringType)
-          .optional()
-          .setDocComment(ts.docComment('Overwrites the datasource url from your schema.prisma file')),
-      )
-      .add(
-        ts
-          .property('errorFormat', ts.namedType('ErrorFormat'))
-          .optional()
-          .setDocComment(ts.docComment('@default "colorless"')),
-      )
-      .add(
-        ts.property('log', ts.array(ts.unionType([ts.namedType('LogLevel'), ts.namedType('LogDefinition')]))).optional()
-          .setDocComment(ts.docComment`
-             @example
-             \`\`\`
-             // Defaults to stdout
-             log: ['query', 'info', 'warn', 'error']
-
-             // Emit as events
-             log: [
-               { emit: 'stdout', level: 'query' },
-               { emit: 'stdout', level: 'info' },
-               { emit: 'stdout', level: 'warn' }
-               { emit: 'stdout', level: 'error' }
-             ]
-             \`\`\`
-             Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client/logging#the-log-option).
-          `),
-      )
-
-    const transactionOptions = ts
-      .objectType()
-      .add(ts.property('maxWait', ts.numberType).optional())
-      .add(ts.property('timeout', ts.numberType).optional())
-
-    if (this.context.dmmf.hasEnumInNamespace('TransactionIsolationLevel', 'prisma')) {
-      transactionOptions.add(ts.property('isolationLevel', ts.namedType('TransactionIsolationLevel')).optional())
+  private applyPendingMigrationsDefinition(this: PrismaClientClass) {
+    if (this.runtimeName !== 'react-native') {
+      return null
     }
 
-    clientOptions.add(
-      ts.property('transactionOptions', transactionOptions).optional().setDocComment(ts.docComment`
-             The default values for transactionOptions
-             maxWait ?= 2000
-             timeout ?= 5000
-          `),
-    )
-
-    if (['library', 'client'].includes(this.runtimeName) && this.context.isPreviewFeatureOn('driverAdapters')) {
-      clientOptions.add(
-        ts
-          .property('adapter', ts.unionType([ts.namedType('runtime.SqlDriverAdapterFactory'), ts.namedType('null')]))
-          .optional()
-          .setDocComment(
-            ts.docComment('Instance of a Driver Adapter, e.g., like one provided by `@prisma/adapter-planetscale`'),
-          ),
+    const method = ts
+      .method('$applyPendingMigrations')
+      .setReturnType(tsx.promise(ts.voidType))
+      .setDocComment(
+        ts.docComment`Tries to apply pending migrations one by one. If a migration fails to apply, the function will stop and throw an error. You are responsible for informing the user and possibly blocking the app as we cannot guarantee the state of the database.`,
       )
-    }
 
-    clientOptions.add(
-      ts.property('omit', ts.namedType('GlobalOmitConfig')).optional().setDocComment(ts.docComment`
-        Global configuration for omitting model fields by default.
-
-        @example
-        \`\`\`
-        const prisma = new PrismaClient({
-          omit: {
-            user: {
-              password: true
-            }
-          }
-        })
-        \`\`\`
-      `),
-    )
-
-    return clientOptions
+    return ts.stringify(method, { indentLevel: 1, newLine: 'leading' })
   }
 }
