+++ packages/client-generator-ts/src/TSClient/file-generators/ModelsFile.ts
@@ -1,10 +1,14 @@
-import type { FileMap } from '../../generateClient'
 import { GenerateContext } from '../GenerateContext'
 
-export function createModelsFile(context: GenerateContext, modelsFileMap: FileMap): string {
-  return `
-${Object.keys(modelsFileMap)
-  .map((m) => `export type * from './models/${m}'`)
-  .join('\n')}
+const jsDocHeader = `/**
+ * This is a barrel export file for all models and their related types.
+ *
+ * 🟢 You can import this file directly.
+ */
 `
+
+export function createModelsFile(context: GenerateContext, modelsNames: string[]): string {
+  const exports = modelsNames.map((m) => `export type * from './models/${context.importFileName(m)}'`)
+  exports.push(`export type * from './${context.importFileName('commonInputTypes')}'`)
+  return jsDocHeader + exports.join('\n')
 }
