+++ packages/client-generator-ts/src/TSClient/index.ts
@@ -1,4 +1,3 @@
 export { Enum } from './Enum'
 export { InputField, InputType } from './Input'
-export { Model, ModelDelegate } from './Model'
 export { TSClient } from './TSClient'
